chore: resolve deprecations in tests
