+++ jadx-core/src/test/java/jadx/tests/integration/others/TestN21.java
@@ -1,0 +1,17 @@
+package jadx.tests.integration.others;
+
+import org.junit.jupiter.api.Test;
+
+import jadx.tests.api.SmaliTest;
+
+import static jadx.tests.api.utils.assertj.JadxAssertions.assertThat;
+
+public class TestN21 extends SmaliTest {
+
+	@Test
+	public void test() {
+		assertThat(getClassNodeFromSmali())
+				.code()
+				.countString(2, "while (");
+	}
+}
