+++ jadx-core/src/test/java/jadx/tests/integration/others/TestReturnWrapping.java
@@ -1,4 +1,4 @@
-package jadx.tests.integration;
+package jadx.tests.integration.others;
 
 import org.junit.jupiter.api.Test;
